requirements.txt
APScheduler==3.10.1
Flask==2.2.3
Flask_Cors==3.0.10
flask_restx==1.1.0
geojson==3.0.1
geopandas==0.12.2
pandas==1.5.3
psycopg2==2.9.5
pyaurn==0.1.21
pytest==7.2.2
python-dotenv==1.0.0
requests==2.28.2
